requirements.txt
# Pvragon Activity Tracker - Python Dependencies
# ==============================================

# Google APIs
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
gspread>=5.0.0

# HTTP requests
requests>=2.28.0

# Data processing
pandas>=2.0.0

# Timezone handling
pytz>=2023.0

# Environment variables
python-dotenv>=1.0.0
